+++ src/Credfeto.DotNet.Repo.Tools.Dependencies/ITrackingHashGenerator.cs
@@ -1,0 +1,10 @@
+using System.Threading;
+using System.Threading.Tasks;
+using Credfeto.DotNet.Repo.Tools.Models;
+
+namespace Credfeto.DotNet.Repo.Tools.Dependencies;
+
+public interface ITrackingHashGenerator
+{
+    ValueTask<string> GenerateTrackingHashAsync(RepoContext repoContext, CancellationToken cancellationToken);
+}